ref(js): Move and code split admin views
