+++ src/sentry/static/sentry/app/views/admin/adminBuffer.jsx
